Be consistent with the first one. A macro might be a good idea for forcing
an integer result to a BOOL.

While, I'm 'OK' with the previous check-in, I'm leary of forcing
all BOOL to be either TRUE or FALSE. We have 'if (f->flag & MASK)'
constructs all over the place, and I don't want to change them all and make
them less readable than they already are.

However, I'm ok with it when the value is stored in a BOOL because unless
you force it to a TRUE or FALSE value, you are assuming that BOOL is an
integer. What if BOOL is defined as a char or as a bit?