TinyMUSH 2.0 bug: exits & rooms can go home. Patch from Alierak.
Alierak:

I'm not sure if TinyMUSH 1.0 isn't to blame instead, but I don't have
TinyMUSH 1.0 to refer to.  There are odd consequences of the ability of
exits & rooms to use the home command.  We probably don't want them to
be able to do that.  The patch below is for TinyMUX 2.2, and causes the
command parser to skip the "home" check if the executor can't have a
location (similar to skipping the checks for exit names).