After a page split, the original page was being marked as
HF_CACHE_UNPROTECTED. This would leave cache entries covering the same
offset in the netmux.pag file. The older one might have 'aged out' before
the newer one. And, the older one would contain all the records of the
newer one and more, but this is still a sobering bug.