Second portion for Issue 256 - Review mail in folder w/o changing folders.
Allow @mail/list <folder>=<msg-list>.   <msg-list> must be present or alternatively an asterick character to allow folder viewing.

Too many hoops are needed to get you to this point to make this clean.   It feels like the hack it is but there isn't a clean path through the mail code to make the capability possible currently.

git-svn-id: https://tinymux.googlecode.com/svn/branches/dev_ronan@2304 d1b986fa-651c-0410-a323-35a8662cf44d