Fix compiler warnings for freeing (const char*)
pValidName is a const char*, so under alternative mux allocators (yes,
yes I know) this can produce warnings when things try to free memory
using it.

The original value of pValidName is not freed, as it really is (const
char*), and munge_space allocates a new LBUF for its munging.  s_Name
duplicates its input, so really all that needs freeing is the output of
munge_space, which the original code did.  It's the casting from char*
to const char* that's problematic because it makes the free (via
pValidName) illegal under more direct allocator methods.

c.f. line 379 under the TYPE_PLAYER case, which handles this right--
using buff just like in this patch.